backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
langchain
langchain-community
ollama
xinference-client
